View files has been modified via AbstractView class
