+++ src/view/card-view.js
@@ -1,5 +1,5 @@
 import {START_INDEX} from '../const';
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view';
 
 const getCommentLength = (comments) => {
   const length = comments.length;
@@ -63,19 +63,12 @@ const createCardTemplate = (movie) => {
     </article>`;
 };
 
-export default class CardView {
-  #element = null;
+export default class CardView extends AbstractView {
   #movie = null;
 
   constructor(movie) {
+    super();
     this.#movie = movie;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
   }
 
   get template() {
@@ -83,11 +76,7 @@ export default class CardView {
   }
 
   get link() {
-    return this.#element.querySelector('.film-card__link');
-  }
-
-  removeElement() {
-    this.#element = null;
+    return this.element.querySelector('.film-card__link');
   }
 
 }
