+++ src/view/filters-view.js
@@ -1,4 +1,5 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view';
+
 
 const FILTER = {
   all: 'All movies',
@@ -36,27 +37,16 @@ const createFiltersTemplate = (filters) => {
       </nav>`;
 };
 
-export default class FiltersView {
-  #element = null;
+export default class FiltersView extends AbstractView {
   #filters = null;
 
   constructor(filters) {
+    super();
     this.#filters = filters;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
   }
 
   get template() {
     return createFiltersTemplate(this.#filters);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 
 }
