+++ src/view/list-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view.js';
 
 const createListTemplate = () => (
   `<section class="films">
@@ -9,21 +9,10 @@ const createListTemplate = () => (
     </section>`
 );
 
-export default class ListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
+export default class ListView extends AbstractView {
 
   get template() {
     return createListTemplate();
   }
 
-  removeElement() {
-    this.#element = null;
-  }
 }
