+++ src/view/sort-view.js
@@ -1,4 +1,5 @@
-import {createElement} from '../render.js';
+import AbstractView from './abstract-view';
+
 
 const createSortTemplate = () => (
   `<ul class="sort">
@@ -8,22 +9,10 @@ const createSortTemplate = () => (
     </ul>`
 );
 
-export default class SortView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-    return this.#element;
-  }
+export default class SortView extends AbstractView {
 
   get template() {
     return createSortTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 
 }
